+++ src/UrlHandlers/UrlHandler.php
@@ -296,11 +296,7 @@ abstract class UrlHandler implements GeneratesResponseInterface
         $context = new PhpContext();
         $context->UrlHandler = $this;
 
-        $this->onBeforeUrlFragmentMatched($request, $currentUrlFragment);
-
         $this->matchingUrl = $this->getMatchingUrlFragment($request, $currentUrlFragment);
-
-        $this->onAfterUrlFragmentMatched($request, $currentUrlFragment);
 
         if ($this->parentHandler) {
             $this->handledUrl = $this->parentHandler->handledUrl . $this->matchingUrl;
@@ -363,6 +359,8 @@ abstract class UrlHandler implements GeneratesResponseInterface
         }
 
         if (stripos($currentUrlFragment, $this->url) === 0) {
+            $this->onUrlFragmentMatched($request, $currentUrlFragment);
+
             return $this->url;
         }
 
@@ -370,23 +368,12 @@ abstract class UrlHandler implements GeneratesResponseInterface
     }
 
     /**
-     * Opportunity to add additional logic before getMatchingUrlFragment is executed
+     * Opportunity to add additional logic if the URL fragment is matched
      *
      * @param Request $request
      * @param string $currentUrlFragment
      */
-    protected function onBeforeUrlFragmentMatched(Request $request, $currentUrlFragment = "")
-    {
-
-    }
-
-    /**
-     * Opportunity to add additional logic after getMatchingUrlFragment is executed
-     *
-     * @param Request $request
-     * @param string $currentUrlFragment
-     */
-    protected function onAfterUrlFragmentMatched(Request $request, $currentUrlFragment = "")
+    protected function onUrlFragmentMatched(Request $request, $currentUrlFragment = "")
     {
 
     }