+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ### 1.6.13
 
-* Added:   Added functions onBeforeUrlFragmentMatched() and onAfterUrlFragmentMatched() to allow an opportunity to add custom
-           logic before a matching URL is found from a URL Handler
+* Added:   Added function onUrlFragmentMatched() to offer and opportunity to augment behaviour of the URL Handler
 
 ### 1.6.12